Introduce Result::cancelled() and Result::wasCancelled() to track user aborts. Factor out ResultData from Result.
