+++ RoboFile.php
@@ -1,5 +1,6 @@
 <?php
 use Symfony\Component\Finder\Finder;
+use Robo\Result;
 
 class RoboFile extends \Robo\Tasks
 {
@@ -296,7 +297,7 @@ class RoboFile extends \Robo\Tasks
         $this->say('You are '.$answer);
         $yes = $this->confirm('Do you want one more question?');
         if (!$yes) {
-            return;
+            return Result::cancelled();
         }
         $lang = $this->askDefault('what is your favorite scripting language?', 'PHP');
         $this->say($lang);
